requirements.txt
backports.shutil-get-terminal-size==1.0.0
decorator==4.0.10
ipython==5.0.0
ipython-genutils==0.1.0
pathlib2==2.1.0
pexpect==4.2.0
pickleshare==0.7.2
prompt-toolkit==1.0.3
ptyprocess==0.5.1
Pygments==2.1.3
PyYAML==3.11
selenium==2.53.6
simplegeneric==0.8.1
six==1.10.0
traitlets==4.2.2
wcwidth==0.1.7
